refactor(chatbot): modularize assistant engine and harden moderation flow
