+++ SortVision/.env.example
@@ -35,9 +35,15 @@ NEXT_PUBLIC_ENABLE_API_LOGGING=true
 # ===========================================
 # Chatbot Configuration
 # ===========================================
-# Gemini API key for AI chatbot functionality
-# Get your key from: https://makersuite.google.com/app/apikey
-NEXT_PUBLIC_GEMINI_API_KEY=your_gemini_api_key_here
+# Chat provider: local (free) or nvidia (remote)
+NEXT_PUBLIC_CHAT_PROVIDER=local
 
-# Gemini API endpoint (optional, defaults to /api/gemini)
-# NEXT_PUBLIC_GEMINI_ENDPOINT=/api/gemini
+# NVIDIA OpenAI-compatible API key (required only for remote AI mode)
+NVIDIA_API_KEY=your_nvidia_api_key_here
+NVIDIA_BASE_URL=https://integrate.api.nvidia.com/v1
+NVIDIA_MODEL=moonshotai/kimi-k2-instruct
+NVIDIA_MODEL_FALLBACKS=
+NVIDIA_TIMEOUT_MS=15000
+NVIDIA_TEMPERATURE=0.6
+NVIDIA_TOP_P=0.9
+NVIDIA_MAX_TOKENS=1024
