+++ SortVision/package.json
@@ -41,6 +41,7 @@
     "@types/react-dom": "^19.2.3",
     "@vercel/analytics": "^2.0.1",
     "@vercel/speed-insights": "^2.0.0",
+    "bad-words": "^4.0.0",
     "braces": "^3.0.3",
     "canvas-confetti": "^1.9.4",
     "class-variance-authority": "^0.7.1",
@@ -59,6 +60,7 @@
     "node-fetch": "^3.3.2",
     "node-html-parser": "^7.1.0",
     "on-headers": "^1.1.0",
+    "openai": "^6.34.0",
     "postcss": "^8.5.9",
     "react": "^19.2.5",
     "react-dom": "^19.2.5",
